Announce /clan joins, add MemePackManager
Since we decided to go back to sticking music in the resource pack (size
of pack makes no difference regarding how fast it loads up, after it's
downloaded of course)

+++ src/main/java/me/robomwm/MountainDewritoes/Music/MusicThing.java
@@ -16,7 +16,7 @@ class MusicThing
         this(URL, seconds, false);
     }
 
-    public MusicThing(String URLorName, int seconds, boolean isName)
+    public MusicThing(String URLorName, long seconds, boolean isName)
     {
         if (isName)
             this.soundName = URLorName;
